TNDO-2645 Rename path-helper.ts to path.helpers.ts
